refactor(backend): DTO 파일을 modules/projects/dto/로 이동
backend/dtos/ 하위에 있던 모든 DTO 파일을 사용처인
modules/projects/dto/로 이동하고 import 경로를 갱신했다.
