+++ backend/modules/projects/dto/projectGitInfoDto.ts
@@ -1,6 +1,6 @@
 import { filter, map, pipe, sortBy, uniqueBy } from 'remeda';
 
-import type { ProjectGitRemote } from '../types/projects.js';
+import type { ProjectGitRemote } from '../../../types/projects.js';
 
 export class ProjectGitRemoteDto implements ProjectGitRemote {
   name: string;
