+++ backend/modules/projects/dto/pullRequestsDto.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { PRListItemDto } from './pullRequestsDto.js';
-import type { GraphQLPRNode } from '../types/pullRequests.js';
+import type { GraphQLPRNode } from '../../../types/pullRequests.js';
 
 describe('PRListItemDto.fromGraphQL', () => {
   const baseNode: GraphQLPRNode = {
